third_party/WebKit/LayoutTests/media/encrypted-media/encrypted-media-lifetime-mediakeys-with-session.html

<!DOCTYPE html>
<html>
    <head>
        <title>Test MediaKeys lifetime when adding a session</title>
        <script src="encrypted-media-utils.js"></script>
        <script src="../../resources/testharness.js"></script>
        <script src="../../resources/testharnessreport.js"></script>
    </head>
    <body>
        <div id="log"></div>
        <script>
            // Since MediaKeys are not ActiveDOMObjects, it is hard to
            // determine when they are garbage collected. For the test below
            // a MediaKeySession (which is an ActiveDOMObject) is added so
            // there is something to count.

            // MediaKeySessions remain as long as:
            //   JavaScript has a reference to it
            //   OR (MediaKeys is around
            //       AND the session has not received a close() event)
            // In the tests below, we do not close any session nor keep a
            // Javascript reference to any session, so MediaKeySessions remain
            // as long as the associated MediaKeys object is around.

            // For this test, create a MediaKeySession (which is an
            // ActiveDOMObject) and verify lifetime.
            async_test(function(test)
            {
                var initDataType;
                var initData;
                var mediaKeys;
                var startingActiveDOMObjectCount = window.internals.activeDOMObjectCount(document);

                function numActiveDOMObjectsCreated()
                {
                    return window.internals.activeDOMObjectCount(document) - startingActiveDOMObjectCount;
                }

                // Create a MediaKeys object with a session.
                getSupportedInitDataType().then(function(type) {
                    initDataType = type;
                    initData = getInitData(initDataType);
                    return navigator.requestMediaKeySystemAccess('org.w3.clearkey', [{}]);
                }).then(function(access) {
                    assert_equals(access.keySystem, 'org.w3.clearkey');
                    return access.createMediaKeys();
                }).then(function(result) {
                    mediaKeys = result;

                    // Verify MediaKeys are not an ActiveDOMObject.
                    // In non-Oilpan, numActiveDOMObjectsCreate() == 0.
                    // In Oilpan, numActiveDOMObjectsCreate() <= 2.
                    // (1 MediaKeysInitializer and
                    //  1 MediaKeySystemAccessInitializer).
                    assert_less_than_equal(numActiveDOMObjectsCreated(), 2, 'MediaKeys.create()');

                    var mediaKeySession = mediaKeys.createSession();
                    return mediaKeySession.generateRequest(initDataType, initData);
                }).then(function() {
                    // 1 MediaKeySession.
                    // In non-Oilpan, numActiveDOMObjectsCreate() == 1.
                    // In Oilpan, numActiveDOMObjectsCreate() <= 4.
                    // (1 MediaKeysInitializer,
                    //  1 MediaKeySystemAccessInitializer,
                    //  1 ContentDecryptionModuleResultPromise and
                    //  1 MediaKeySession).
                    assert_less_than_equal(numActiveDOMObjectsCreated(), 4, 'MediaKeys.createSession(1)');

                    // Run gc(), should not affect MediaKeys object nor the
                    // session since we still have a reference to it.

                    // When enabling oilpan GC, the in-active
                    // ScriptPromiseResolvers will be destroyed.
                    return createGCPromise();
                }).then(function(result) {
                    assert_equals(typeof mediaKeys.createSession, 'function');
                    // Ensure that MediaKeySession (but not PromiseResolver) is
                    // still around.
                    assert_less_than_equal(numActiveDOMObjectsCreated(), 2, 'After gc()');

                    // Drop reference to the MediaKeys object and run gc()
                    // again. Object should be collected this time. Since
                    // MediaKeySessions remain alive as long as MediaKeys is
                    // around, it is possible that gc() checks the
                    // MediaKeySession object first, and doesn't collect it
                    // since MediaKeys hasn't been collected yet. Thus run gc()
                    // twice, to ensure that the unreferenced MediaKeySession
                    // object get collected.
                    mediaKeys = null;
                    return createGCPromise();
                }).then(function(result) {
                    return createGCPromise();
                }).then(function(result) {
                    assert_less_than_equal(numActiveDOMObjectsCreated(), 1);
                    test.done();
                }).catch(function(error) {
                    forceTestFailureFromPromise(test, error);
                });
            }, 'MediaKeys lifetime with session');
        </script>
    </body>
</html>
